tweaked some positions, added a second test milk bottle

+++ Assets/Scripts/Hud.cs
@@ -9,6 +9,6 @@ public class Hud : MonoBehaviour
 
     void Update()
     {
-        EnergyText.text = GameState.Energy.ToString();
+        EnergyText.text = ((int)GameState.Energy).ToString();
     }
 }
